Create and Edit methods
These show the  blogPhoto controller methods for create and edit pages.  They both take an image file as an argument from the view. Call on the convertPhoto() method to change the image file to a byte[] to be stored in the database.  It also makes use of a bind to map the http request data with the blogPhoto model and sets a blogPhoto object as a parameter.  Once all the data is in the proper form, it is then saved to the data base as a member of the blogPhoto table.