byte[] to string conversion
This converts a byte[] from the database's blogPhoto table to base 64 string from within the index view to be displayed for the user on the index page. This snippet utilizes viewbag within the razor page to be called on in the html image tag as the source.  The image tag is nested within an anchor tag to create a clickable link out of the image being displayed that sends the user to the details page.