Index page
This page was styled using a combination of bootstrap and external CSS.  It displays the blogPhotos from the database.  when hovered over with the mouse, the image expands and the edit and delete buttons appear in the lower left corner of the image.  When the image itself is clicked on the user is sent to the image details page.